pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gitignore"
version = "1.3.0"
description = "Creates a basic .gitignore file in the current directory"
requires-python = ">=3.11"
dependencies = [
    "platformdirs",
    "PyYAML",
]

[project.scripts]
gitignore = "gitignore.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gitignore = ["config.yaml"]
